feat: finsihed map generation
