+++ gb/public/scripts/index.mjs
@@ -2,6 +2,9 @@ const data = await (await fetch("../data/index.json")).json();
 const areas = await (await fetch("../data/primary-area.json")).json();
 
 import { setSearchEventListeners, setSortingEventListeners, setUpAllPrimaryAreas } from "./table.mjs";
+import {  tooltip } from "./map.mjs";
+
+tooltip();
 
 setUpAllPrimaryAreas(data);
