+++ pipelines/gb/configuration.mjs
@@ -3,4 +3,11 @@ export const config = {
   propertyPrices: {
     fetch: false,
   },
+  indexTitle: "GB",
+  scales: {
+    gothenburgMedianIncomeToMedianHousePrice: 0.25,
+    ownershipRate: 0.25,
+    netMigration: 0.25,
+    overCrowdingRate: 0.25,
+  }
 };
